Use db in persistent dir (#44)
* Add more tests, make log level configurable

* Add Social Presence module, allow throwing in mapping. Fix Adapters accordingly.

* Fix namespaces not being supported with erasableSyntaxOnly

* Fix schema issues, allow nullable puuid resolver values to indicate explicit failure

* Restructure async behavior, Let match history manager only react on (sucessfully) downloaded matches.

* WIP, use a db instead of raw json files

* More changes

+++ backend/shared/schemas/StorageStatusDTO.ts
@@ -2,8 +2,8 @@ import { z } from "zod";
 
 export const StorageStatusDTOSchema = z.object({
     isSetup: z.boolean(),
-    matchCount: z.number(),
-    totalSizeBytes: z.number(),
+    matchCount: z.number().int(),
+    totalSizeBytes: z.number().int(),
 });
 
 export type StorageStatusDTO = z.infer<typeof StorageStatusDTOSchema>;
